Bug fixes
The mask wasn't correctly calculated, as a mask of 0 would lead to a left shift
of 32, which on x86 will not be performed (Only the lowest 5 bit will be used,
which are 0 in that case).
The destination address wasn't changed for the `sock_addr` structure handed to
`send_to` and therefore the packed wouldn't be sent to the correct address.
For RC4 the crypto library has to be linked against. This is reflected now in the
setup file.